combined the prover proposal scripts
